Re-wrote save command and moved it to command file
