+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "check-word": "^1.1.0",
     "mongodb": "^4.13.0",
+    "prompt-sync": "^4.2.0",
     "vorpal": "^1.12.0"
   }
 }
