+++ src/classes/Commands.js
@@ -2,8 +2,13 @@
 const wordsClass = require("check-word");
 const Database = require("./Database");
 
+const prompt = require('prompt-sync')();
+
 // dictionary object used to check whether a word is valid or not
 const dictionary = wordsClass("en");
+
+//file system module
+var fs = require("fs");
 
 /**
  * Commands class used to store static helper methods for the cli.
@@ -76,7 +81,7 @@ class Commands {
   static async newPuzzle(GameManager, Database) {
     if (GameManager.isPuzzleOpen) {
       console.log("game is in progess");
-      //promptSave();
+      this.promptSave(GameManager);
       return;
     }
 
@@ -192,6 +197,46 @@ class Commands {
 
     console.log(GameManager.foundWords);
   }
-}
 
+  /**
+   * Saves current puzzle
+   * @param {GameManager} GameManager - object used to keep track of the game/player
+   * @param {fileName} fileName - users inputted file name
+   * @returns null
+   */
+  static save(fileName, GameManager){
+    if(GameManager.isPuzzleOpen == false)
+    {
+      console.log("SpellingBee> No puzzle open, you can not save");
+      return;
+    }
+    else
+    {
+      let table = {
+        words: GameManager.foundWords,
+        pangram: GameManager.pangram,
+        requiredLetter: GameManager.requiredLetter,
+        userPoints: GameManager.userPoints,
+      };
+      let jsonFile = JSON.stringify(table);
+      fs.writeFile(fileName + ".json", jsonFile, 'utf8', (err) => {if (err) throw err;});
+      console.log('SpellingBee> The file has been saved!');
+      GameManager.isPuzzleOpen = false;
+      return;
+    }
+  }
+
+  static promptSave(GameManager){
+    let save = prompt('SpellingBee> Would you like to save your current game? (y/n) ');
+    if(save == 'y'){
+      let fileName = prompt('SpellingBee> Enter a file name: ');
+      this.save(fileName, GameManager);
+    }
+    else{
+      console.log("SpellingBee> Game not saved");
+      GameManager.isPuzzleOpen = false;
+      return;
+    }
+  }
+}
 module.exports = Commands;
